feat: enhance BudgetLineItemService with user association checks and refactor CRUD operations
