+++ backend/ops_api/ops/services/ops_service.py
@@ -12,7 +12,7 @@ class OpsService(Protocol[T]):
 
     def get(self, id: int) -> T: ...
 
-    def get_list(self, search: Optional[str] = None) -> list[T]: ...
+    def get_list(self, data: dict | None) -> tuple[list[T], dict | None]: ...
 
 
 class ServiceError(Exception):
